seat: split into multiple files
