+++ types/meson.build
@@ -5,13 +5,16 @@ lib_wlr_types = static_library(
 		'data_device/wlr_data_offer.c',
 		'data_device/wlr_data_source.c',
 		'data_device/wlr_drag.c',
+		'seat/wlr_seat_keyboard.c',
+		'seat/wlr_seat_pointer.c',
+		'seat/wlr_seat_touch.c',
+		'seat/wlr_seat.c',
 		'wlr_box.c',
 		'wlr_compositor.c',
 		'wlr_cursor.c',
 		'wlr_gamma_control.c',
 		'wlr_idle_inhibit_v1.c',
 		'wlr_idle.c',
-		'wlr_idle_inhibit_v1.c',
 		'wlr_input_device.c',
 		'wlr_input_inhibitor.c',
 		'wlr_keyboard.c',
@@ -26,7 +29,6 @@ lib_wlr_types = static_library(
 		'wlr_primary_selection.c',
 		'wlr_region.c',
 		'wlr_screenshooter.c',
-		'wlr_seat.c',
 		'wlr_server_decoration.c',
 		'wlr_surface.c',
 		'wlr_tablet_pad.c',
@@ -34,9 +36,9 @@ lib_wlr_types = static_library(
 		'wlr_touch.c',
 		'wlr_wl_shell.c',
 		'wlr_xcursor_manager.c',
+		'wlr_xdg_output.c',
 		'wlr_xdg_shell_v6.c',
 		'wlr_xdg_shell.c',
-		'wlr_xdg_output.c',
 	),
 	include_directories: wlr_inc,
 	dependencies: [pixman, xkbcommon, wayland_server, wlr_protos],
